// src/Flip/Client/generated/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type ClientOptions = {
    baseUrl: 'http://localhost:23901/' | (string & {});
};

export type ChangeEntityTypeModel = {
    unique: string;
    contentTypeUnique: string;
    templateUnique?: null | string;
    properties?: null | Array<EntityTypePropertyModel>;
};

export type ContentTypeModel = {
    name?: null | string;
    unique: string;
    defaultTemplateUnique?: null | string;
    propertyTypes: Array<PropertyTypeModel>;
    allowedTemplates: Array<TemplateModel>;
};

export type EntityTypePropertyModel = {
    label?: null | string;
    alias?: null | string;
    newAlias?: null | string;
    editor?: null | string;
    dataTypeUnique?: null | string;
};

export type ProblemDetails = {
    type?: null | string;
    title?: null | string;
    status?: null | number | string;
    detail?: null | string;
    instance?: null | string;
};

export type PropertyTypeModel = {
    name?: null | string;
    alias?: null | string;
    dataTypeUnique: string;
    propertyEditorAlias?: null | string;
};

export type TemplateModel = {
    name?: null | string;
    unique: string;
};

export type PostChangeTypeData = {
    body: ChangeEntityTypeModel;
    path?: never;
    query?: {
        entityType?: string;
    };
    url: '/umbraco/flip/management/api/v1/change-type';
};

export type PostChangeTypeErrors = {
    /**
     * Bad Request
     */
    400: ProblemDetails;
    /**
     * The resource is protected and requires an authentication token
     */
    401: unknown;
};

export type PostChangeTypeError = PostChangeTypeErrors[keyof PostChangeTypeErrors];

export type PostChangeTypeResponses = {
    /**
     * OK
     */
    200: unknown;
};

export type GetContentModelData = {
    body?: never;
    path?: never;
    query?: {
        unique?: string;
        entityType?: string;
    };
    url: '/umbraco/flip/management/api/v1/content-model';
};

export type GetContentModelErrors = {
    /**
     * The resource is protected and requires an authentication token
     */
    401: unknown;
};

export type GetContentModelResponses = {
    /**
     * OK
     */
    200: ChangeEntityTypeModel;
};

export type GetContentModelResponse = GetContentModelResponses[keyof GetContentModelResponses];

export type GetPermittedData = {
    body?: never;
    path?: never;
    query?: {
        unique?: string;
        entityType?: string;
    };
    url: '/umbraco/flip/management/api/v1/permitted';
};

export type GetPermittedErrors = {
    /**
     * The resource is protected and requires an authentication token
     */
    401: unknown;
};

export type GetPermittedResponses = {
    /**
     * OK
     */
    200: Array<ContentTypeModel>;
};

export type GetPermittedResponse = GetPermittedResponses[keyof GetPermittedResponses];